:recycle: Move Responses into own module
